+++ internal/httpapi/assets/_app/immutable/chunks/I4roHWlF.js
@@ -1,0 +1,1 @@
+import{b as f,h as o,c as v,E as y,r as l,s as m,d as p,e as i,H as T,f as A}from"./BkxjBBuX.js";import{B as E}from"./BPAE2Pgv.js";function g(d,_,e){var s;o&&(s=A,v());var r=new E(d);f(()=>{var a=_()??null;if(o){var h=l(s),c=h===T,u=a!==null;if(c!==u){var t=m();p(t),r.anchor=t,i(!1),r.ensure(a,a&&(n=>e(n,a))),i(!0);return}}r.ensure(a,a&&(n=>e(n,a)))},y)}export{g as c};